requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
openai
python-dotenv
python-docx
matplotlib
pywin32